requirements.txt
tensorflow_gpu==2.3.0
numpy==1.18.5
opencv_python==4.1.0.25
matplotlib==3.0.3
tensorflow_hub==0.7.0
tqdm==4.31.1
Pillow==8.0.1
tensorflow==2.3.1
